requirements.txt
asgiref==3.3.4
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-formtools==2.3
django-heroku==0.3.1
django-otp==1.0.3
django-phonenumber-field==5.1.0
django-two-factor-auth==1.13.1
djangorestframework==3.12.4
gunicorn==20.1.0
joblib==1.0.1
numpy==1.20.2
phonenumbers==8.12.21
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
qrcode==6.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
whitenoise==5.2.0
